Optimized and fixed table controller

+++ app/Http/Controllers/Admin/TableCtrl.php
@@ -8,8 +8,6 @@ use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
-use Intervention\Image\Facades\Image;
-use Symfony\Component\HttpFoundation\File\UploadedFile;
 
 class TableCtrl extends Controller
 {
@@ -25,41 +23,21 @@ class TableCtrl extends Controller
     function __construct()
     {
 
-        // get model name from the URL
+        $this->modelName = UrlCtrl::getModelNameFromUrl();
 
-        $urlParts = explode('/', url()->current());
-        $key = array_search('_t_', $urlParts);
-        $this->modelName = array_get($urlParts, $key + 1);
+        $this->tableName = UrlCtrl::getTableNameFromModelName($this->modelName);
 
-        //get table name by model name
-
-        $this->tablesJson = json_decode(Storage::disk('table_config')->get('tables.json'), true);
-        foreach ($this->tablesJson as $tableName => $tableMeta) {
-            if ($tableMeta['modelName'] == $this->modelName) {
-                $this->tableName = $tableName;
-            }
-        }
-        if (!$this->tableName) {
-            abort(404);
-        }
+        $this->tablesJson = FileCtrl::getFile('table_config', 'tables.json', true);
 
         $this->tableMeta = $this->tablesJson[$this->tableName];
 
-        $this->columnsMeta = json_decode(Storage::disk('table_config')->get($this->tableName . '_columns.json'), true);
+        $this->middleware('checkUserLevel:' . $this->tableMeta['userLevel']);
+
+        $this->columnsMeta = FileCtrl::getFile('table_config', $this->tableName . '_columns.json', true);
 
         $this->recordsUserFilter = $this->tableMeta['recordsUserFilter'];
 
-        // Construct initial data
-
-        $this->initData = [
-            'modelIndexUrl' => $this->urlPrefix . $this->modelName . '/',
-            'modelName' => $this->modelName,
-            'pageTitle' => $this->tableMeta['pageTitle'],
-            'indexRecordsUrl' => $this->urlPrefix . $this->modelName . '/get' . $this->modelName . 'Records',
-            'imageDir' => '/images/' . $this->modelName . '/',
-            'imageThumbDir' => '/images/' . $this->modelName . '/thumb/',
-            'fileDir' => '/files/' . $this->modelName . '/'
-        ];
+        $this->initData = DataCtrl::getInitData($this->urlPrefix, $this->modelName, $this->tableMeta);
 
     }
 
@@ -97,55 +75,9 @@ class TableCtrl extends Controller
             $query->take($length);
         }
 
+        $query = DbCtrl::applyJoins($query, $this->tableMeta, $this->tableName);
 
-        // Don't change till here
-
-
-        /*********************
-         *  Need to change this
-         */
-
-        if (isset($this->tableMeta['hasJoins']) && $this->tableMeta['hasJoins']) {
-
-            $joins = $this->tableMeta['joins'];
-
-            $query->select($this->tableName . '.*');
-
-            foreach ($joins as $join) {
-
-                $query->leftJoin(
-                    $join['foreignTable'],
-                    $this->tableName . '.' . $join['localColumn'],
-                    $join['condition'],
-                    $join['foreignTable'] . '.' . $join['foreignColumn']
-                );
-
-                $query->addSelect(
-                    $join['foreignTable']. '.' .
-                    $join['foreignTitleColumn'] .
-                    ' as '.
-                    '_' . $join['foreignTable'] . '_title'
-                );
-
-            }
-
-        }
-
-        /*********************
-         *  change till here
-         */
-
-        // Don't change from here
-
-        // Filter records according to the level of user
-        if ($this->recordsUserFilter && Auth::check()) {
-            $currentUserLevel = Auth::user()->level;
-            $query->leftJoin('users', $this->tableName.'.user_id', '=', 'users.id');
-            $query->addSelect('users.level as _user_level');
-            $query->where('users.level', '>=', $currentUserLevel);
-        }
-
-        // Response Data
+        $query = UserCtrl::filterUserRecords($this->recordsUserFilter, $query, $this->tableName);
 
         $unformattedRecords = $query->get();
 
@@ -161,7 +93,6 @@ class TableCtrl extends Controller
         ];
 
         return $data;
-        // Don't change till here
     }
 
     public function index()
@@ -172,11 +103,16 @@ class TableCtrl extends Controller
 
     public function create()
     {
+
+        $frontRules = DbCtrl::constructFrontRules($this->columnsMeta);
+
         $data = array_merge_recursive($this->initData, [
             'formAction' => $this->urlPrefix . $this->modelName,
             'formType' => 'Add',
-            'columns' => $this->columnsMeta
+            'columns' => $this->columnsMeta,
+            'frontRules' => $frontRules
         ]);
+
         return view('pages.admin.record_form', $data);
     }
 
@@ -197,50 +133,15 @@ class TableCtrl extends Controller
     protected function getSpecificModel($id)
     {
 
-        // Deny access if user tries to retrieve another user level record
-        if ($this->recordsUserFilter && Auth::check()) {
-            $currentUserLevel = Auth::user()->level;
-            $recordUserLevel = DB::table($this->tableName)
-                ->where($this->tableName.'.id', $id)
-                ->leftJoin('users', $this->tableName . '.user_id', '=', 'users.id')
-                ->select('users.*')
-                ->first();
-            if ($currentUserLevel > $recordUserLevel->level) {
-                abort(403, 'Unauthorized action...!');
-            }
-        }
+        UserCtrl::denyUserOnAccess($id, $this->recordsUserFilter, $this->tableName);
 
         $model = null;
 
         $query = DB::table($this->tableName);
 
-        $query->where('student.id', $id);
+        $query->where($this->tableName . '.id', $id);
 
-        if (isset($this->tableMeta['hasJoins']) && $this->tableMeta['hasJoins']) {
-
-            $joins = $this->tableMeta['joins'];
-
-            $query->select($this->tableName . '.*');
-
-            foreach ($joins as $join) {
-
-                $query->leftJoin(
-                    $join['foreignTable'],
-                    $this->tableName . '.' . $join['localColumn'],
-                    $join['condition'],
-                    $join['foreignTable'] . '.' . $join['foreignColumn']
-                );
-
-                $query->addSelect(
-                    $join['foreignTable']. '.' .
-                    $join['foreignTitleColumn'] .
-                    ' as '.
-                    '_' . $join['foreignTable'] . '_title'
-                );
-
-            }
-
-        }
+        $query = DbCtrl::applyJoins($query, $this->tableMeta, $this->tableName);
 
         $model = $query->first();
         return $model;
@@ -248,16 +149,10 @@ class TableCtrl extends Controller
 
     public function store(Request $request)
     {
-        $rules = [];
 
-        foreach ($this->columnsMeta as $columnName => $column) {
-            if ($column['backRules']) {
-                $rules = array_add($rules, $columnName, $column['backRules']);
-            }
-        }
-        $this->validate($request, $rules);
+        DbCtrl::validateRequest($request, $this->columnsMeta, $this);
 
-        $new_id = (DB::table($this->tableName)->count() > 0) ? (DB::table($this->tableName)->orderBy('id', 'desc')->get(['id'])->first()->id + 1) : (1);
+        $new_id = DbCtrl::getNewId($this->tableName);
 
         $this->saveRecord($request, $new_id);
 
@@ -270,49 +165,11 @@ class TableCtrl extends Controller
         $insertValues = $request->all();
 
 
-        if(sizeof($this->tableMeta['imageColumns'])){
+        $insertValues =
+            DbCtrl::saveImages($request, $this->tableMeta, $new_id, $this->columnsMeta, $this->initData, $insertValues);
 
-            $imageColumns = $this->tableMeta['imageColumns'];
-
-            foreach ($imageColumns as $imageColumn) {
-
-                $image = $request->file($imageColumn);
-
-                if (is_a($image, UploadedFile::class)) {
-
-                    $img_name = $new_id . '.jpg';
-
-                    $img = Image::make($image);
-
-                    $this->saveImage($img, $img_name,
-                        $this->columnsMeta[$imageColumn]['imageWidth'], 90, $this->initData['imageDir']);
-                    $this->saveImage($img, $img_name, 100, 50, $this->initData['imageThumbDir']);
-
-                    $insertValues[$imageColumn] = $img_name;
-                }
-            }
-        }
-
-        if(sizeof($this->tableMeta['fileColumns'])){
-
-            $fileColumns = $this->tableMeta['fileColumns'];
-
-            foreach ($fileColumns as $fileColumn) {
-
-                $file = $request->file($fileColumn);
-
-                if (is_a($file, UploadedFile::class)) {
-
-                    $ext = $file->getClientOriginalExtension();
-
-                    $fileName = $new_id . '.' . $ext;
-
-                    Storage::put($this->initData['fileDir'] . $fileName, file_get_contents($file));
-
-                    $insertValues[$fileColumn] = $fileName;
-                }
-            }
-        }
+        $insertValues =
+            DbCtrl::saveFiles($request, $this->tableMeta, $new_id, $this->initData, $insertValues);
 
         $insertValues['created_at'] = Carbon::now();
         $insertValues['updated_at'] = Carbon::now();
@@ -327,12 +184,11 @@ class TableCtrl extends Controller
         DB::table($this->tableName)->insert($insertValues);
 
 
-
     }
 
     public function show($id)
     {
-        $columnsValues = (array) $this->getSpecificModel($id);
+        $columnsValues = (array)$this->getSpecificModel($id);
 
         $data = array_merge_recursive(
             $this->initData,
@@ -345,13 +201,16 @@ class TableCtrl extends Controller
 
     public function edit($id)
     {
-        $columnsValues = (array) $this->getSpecificModel($id);
+        $columnsValues = (array)$this->getSpecificModel($id);
+
+        $frontRules = DbCtrl::constructFrontRules($this->columnsMeta);
 
         $data = array_merge_recursive($this->initData, [
             'formAction' => $this->urlPrefix . $this->modelName . '/' . $columnsValues['id'],
             'formType' => 'Update',
             'columns' => $this->columnsMeta,
-            'columnsValues' => $columnsValues
+            'columnsValues' => $columnsValues,
+            'frontRules' => $frontRules
         ]);
         return view('pages.admin.record_form', $data);
     }
@@ -359,19 +218,9 @@ class TableCtrl extends Controller
     public function update(Request $request, $id)
     {
 
-        $rules = [];
-
-        foreach ($this->columnsMeta as $columnName => $column) {
-            if ($column['backRules']) {
-                $rules = array_add($rules, $columnName, $column['backRules']);
-            }
-        }
-
-        $this->validate($request, $rules);
+        DbCtrl::validateRequest($request, $this->columnsMeta, $this);
 
         $model = DB::table($this->tableName)->where('id', $id);
-
-        $new_id = $id;
 
         $updateValues = $request->all();
 
@@ -381,57 +230,18 @@ class TableCtrl extends Controller
 
         $updateValues['updated_at'] = Carbon::now();
 
-        if(sizeof($this->tableMeta['imageColumns'])){
+        $updateValues =
+            DbCtrl::saveImages($request, $this->tableMeta, $id, $this->columnsMeta, $this->initData, $updateValues);
 
-            $imageColumns = $this->tableMeta['imageColumns'];
+        $updateValues =
+            DbCtrl::saveFiles($request, $this->tableMeta, $id, $this->initData, $updateValues);
 
-            foreach ($imageColumns as $imageColumn) {
-
-                if (isset($updateValues[$imageColumn])){
-
-                    $image = $request->file($imageColumn);
-
-                    if (is_a($image, UploadedFile::class)) {
-
-                        $img_name = $new_id . '.jpg';
-
-                        $img = Image::make($image);
-
-                        $this->saveImage($img, $img_name,
-                            $this->columnsMeta[$imageColumn]['imageWidth'], 90, $this->initData['imageDir']);
-                        $this->saveImage($img, $img_name, 100, 50, $this->initData['imageThumbDir']);
-
-                        $updateValues[$imageColumn] = $img_name;
-                    }
-                }
+        // Remove null values
+        foreach ($updateValues as $key => $value) {
+            if ($value == null) {
+                array_pull($updateValues, $key);
             }
         }
-
-
-        if(sizeof($this->tableMeta['fileColumns'])){
-
-            $fileColumns = $this->tableMeta['fileColumns'];
-
-            foreach ($fileColumns as $fileColumn) {
-
-                if (isset($updateValues[$fileColumn])){
-
-                    $file = $request->file($fileColumn);
-
-                    if (is_a($file, UploadedFile::class)) {
-
-                        $ext = $file->getClientOriginalExtension();
-
-                        $fileName = $new_id . '.' . $ext;
-
-                        Storage::put($this->initData['fileDir'] . $fileName, file_get_contents($file));
-
-                        $updateValues[$fileColumn] = $fileName;
-                    }
-                }
-            }
-        }
-
 
         $model->update($updateValues);
 
@@ -441,30 +251,17 @@ class TableCtrl extends Controller
     public function destroy($id)
     {
 
-        // Deny access if user tries to retrieve another user level record
-        if ($this->recordsUserFilter && Auth::check()) {
-            $currentUserLevel = Auth::user()->level;
-            $recordUserLevel = DB::table($this->tableName)
-                ->where($this->tableName.'.id', $id)
-                ->leftJoin('users', $this->tableName . '.user_id', '=', 'users.id')
-                ->select('users.*')
-                ->first();
-            if ($currentUserLevel > $recordUserLevel->level) {
-                abort(403, 'Unauthorized action...!');
-            }
-        }
+        UserCtrl::denyUserOnAccess($id, $this->recordsUserFilter, $this->tableName);
 
+        $model = (array)DB::table($this->tableName)->where('id', $id)->first();
 
-        $model = (array) DB::table($this->tableName)->where('id', $id)->first();
-
-
-        if(sizeof($this->tableMeta['imageColumns'])){
+        if (isset($this->tableMeta['imageColumns']) && sizeof($this->tableMeta['imageColumns'])) {
 
             $imageColumns = $this->tableMeta['imageColumns'];
 
             foreach ($imageColumns as $imageColumn) {
 
-                if (isset($model[$imageColumn])){
+                if (isset($model[$imageColumn])) {
 
                     Storage::delete([
                         $this->initData['imageDir'] . $model[$imageColumn],
@@ -474,14 +271,13 @@ class TableCtrl extends Controller
             }
         }
 
-
-        if(sizeof($this->tableMeta['fileColumns'])){
+        if (isset($this->tableMeta['fileColumns']) && sizeof($this->tableMeta['fileColumns'])) {
 
             $fileColumns = $this->tableMeta['fileColumns'];
 
             foreach ($fileColumns as $fileColumn) {
 
-                if (isset($model[$fileColumn])){
+                if (isset($model[$fileColumn])) {
 
                     Storage::delete([
                         $this->initData['fileDir'] . $model[$fileColumn],
@@ -490,24 +286,9 @@ class TableCtrl extends Controller
             }
         }
 
-        DB::table($this->tableName)->where('id' , $id)->delete();
+        DB::table($this->tableName)->where('id', $id)->delete();
 
         return redirect($this->initData['modelIndexUrl']);
-    }
-
-    protected function saveImage($img, $img_name, $width, $quality, $dir)
-    {
-        /***********************
-         * Don't Change this
-         */
-        $img
-            ->resize($width, null, function ($c) {
-                $c->aspectRatio();
-                $c->upsize();
-            })
-            ->encode('jpg', $quality);
-
-        Storage::put($dir . $img_name, $img);
     }
 
     protected function applyAccessors($columns, $records)
@@ -518,13 +299,13 @@ class TableCtrl extends Controller
         // Collecting dates and radios columns
 
         foreach ($columns as $name => $value) {
-                if ($value['formType'] == 'date') {
-                    array_push($dates, $name);
-                }
-                if ($value['formType'] == 'radio') {
-                    array_push($radios, $name);
-                }
+            if ($value['formType'] == 'date') {
+                array_push($dates, $name);
             }
+            if ($value['formType'] == 'radio') {
+                array_push($radios, $name);
+            }
+        }
 
         // Date formatting
 
@@ -551,7 +332,7 @@ class TableCtrl extends Controller
             foreach ($records as $record) {
                 foreach ($radios as $radioColumn) {
 
-                    $radioValue  = $record->$radioColumn;
+                    $radioValue = $record->$radioColumn;
                     $radioButtons = $columns[$radioColumn]['radioButtonsInForm'];
                     $radioTitle = null;
                     foreach ($radioButtons as $radioButton) {
